# Conflicts:
#	install/emoncms_modules.sh
#	install/emonsd.config.ini
#	install/main.sh
#	update/config.ini
#	update/emoncms_core.sh
#	update/emonpi.sh

+++ update/emoncms_core.sh
@@ -1,5 +1,5 @@
 #!/bin/bash
-source config.ini
+source load_config.sh
 
 echo "-------------------------------------------------------------"
 echo "Update Emoncms Core"
@@ -16,18 +16,21 @@ fi
 # This needs to be run prior to emoncms git pull
 # -----------------------------------------------------------------
 cd
-current_settings_md5="$($openenergymonitor_dir/EmonScripts/common/./md5sum.py $emoncms_www/settings.php)"
-echo "Current settings.php md5: $current_settings_md5"
+echo
+if [ -f $emoncms_www/settings.php ]; then
+  current_settings_md5="$($openenergymonitor_dir/EmonScripts/common/./md5sum.py $emoncms_www/settings.php)"
+  echo "Current settings.php md5: $current_settings_md5"
 
-current_default_settings_md5="$($openenergymonitor_dir/EmonScripts/common/md5sum.py $emoncms_www/default.emonpi.settings.php)"
-echo "Default settings.php md5: $current_default_settings_md5"
+  current_default_settings_md5="$($openenergymonitor_dir/EmonScripts/common/md5sum.py $emoncms_www/default.emonpi.settings.php)"
+  echo "Default settings.php md5: $current_default_settings_md5"
 
-if [ "$current_default_settings_md5" == "$current_settings_md5" ]; then
-  echo "settings.php has NOT been user modifed"
-  settings_unmodified=true
-else
-  echo "settings.php HAS been user modified"
-  settings_unmodified=false
+  if [ "$current_default_settings_md5" == "$current_settings_md5" ]; then
+    echo "settings.php has NOT been user modifed"
+    settings_unmodified=true
+  else
+    echo "settings.php HAS been user modified"
+    settings_unmodified=false
+  fi
 fi
 
 # -----------------------------------------------------------------
@@ -43,29 +46,36 @@ if $changes; then
     echo "- no local changes"
     echo "- running: git pull origin $branch"
     echo
+    git fetch --all --prune
     git pull origin $branch
 else
-    echo "- changes"
+    echo "WARNING local changes Emoncms Core not updated"
+    echo "- git status:"
+    echo
+    git status
+    echo
 fi
 
 # -----------------------------------------------------------------
 # check to see if user has modifed settings.php and if update is need. Auto apply of possible
 # -----------------------------------------------------------------
 echo
-new_default_settings_md5="$($openenergymonitor_dir/EmonScripts/common/md5sum.py $emoncms_www/default.emonpi.settings.php)"
-echo "NEW default settings.php md5: $new_default_settings_md5"
+if [ -f $emoncms_www/settings.php ]; then
+  new_default_settings_md5="$($openenergymonitor_dir/EmonScripts/common/md5sum.py $emoncms_www/default.emonpi.settings.php)"
+  echo "NEW default settings.php md5: $new_default_settings_md5"
 
-# check to see if there is an update waiting for settings.php
-if [ "$new_default_settings_md5" != "$current_default_settings_md5" ]; then
-  echo "Update required to settings.php..."
-  if [ $settings_unmodified == true ]; then
-    sudo cp $emoncms_www/default.emonpi.settings.php $emoncms_www/settings.php
-    echo "settings.php autoupdated"
+  # check to see if there is an update waiting for settings.php
+  if [ "$new_default_settings_md5" != "$current_default_settings_md5" ]; then
+    echo "Update required to settings.php..."
+    if [ $settings_unmodified == true ]; then
+      sudo cp $emoncms_www/default.emonpi.settings.php $emoncms_www/settings.php
+      echo "settings.php autoupdated"
+    else
+      echo "**ERROR: unable to autoupdate settings.php since user changes are present, manual review required**"
+    fi
   else
-    echo "**ERROR: unable to autoupdate settings.php since user changes are present, manual review required**"
+    echo "settings.php not updated"
   fi
-else
-  echo "settings.php not updated"
 fi
 
 #########################################################################################
